server/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
elevenlabs==1.50.5
firecrawl-py==1.15.0
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.4
